requirements.txt
pystray
pillow
keyboard
auto-py-to-exe
